refactor: EB-013 Files pattern modified
Component's files pattern changed
