+++ src/components/Layout/components/Content/content.component.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import { RouterConfig } from 'navigation';
+import { Container } from './content.styles';
+
+export const Content: React.FC = () => {
+    return (
+        <Container>
+            <RouterConfig />
+        </Container>
+    );
+};
