+++ src/components/Layout/components/Content/index.tsx
@@ -1,11 +1,1 @@
-import React from 'react';
-import { RouterConfig } from 'navigation';
-import { Container } from './styles';
-
-export const Content: React.FC = () => {
-    return (
-        <Container>
-            <RouterConfig />
-        </Container>
-    );
-};
+export { Content } from './content.component';
