+++ src/components/Layout/components/Footer/footer.component.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+import { Container } from './footer.styles';
+
+export const Footer: React.FC = () => {
+    return (
+        <Container>
+            <h1>Footer</h1>
+        </Container>
+    );
+};
