+++ src/components/Layout/components/Footer/index.tsx
@@ -1,11 +1,1 @@
-import React from 'react';
-
-import { Container } from './styles';
-
-export const Footer: React.FC = () => {
-    return (
-        <Container>
-            <h1>Footer</h1>
-        </Container>
-    );
-};
+export { Footer } from './footer.component';
