+++ src/components/Layout/components/Header/header.component.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+import { Container } from './header.styles';
+
+export const Header: React.FC = () => {
+    return (
+        <Container>
+            <h1>Header</h1>
+        </Container>
+    );
+};
