+++ src/components/Layout/components/Header/index.tsx
@@ -1,11 +1,1 @@
-import React from 'react';
-
-import { Container } from './styles';
-
-export const Header: React.FC = () => {
-    return (
-        <Container>
-            <h1>Header</h1>
-        </Container>
-    );
-};
+export { Header } from './header.component';
