+++ src/navigation/index.ts
@@ -1,2 +1,2 @@
-export * from 'navigation/router.config';
+export * from 'navigation/router-config';
 export * from 'navigation/router.paths';
