+++ src/state/index.ts
@@ -1,3 +1,3 @@
-export * from 'state/contexts/app-state';
-export * from 'state/reducers';
-export * from 'state/actions';
+export * from './contexts/app-state';
+export * from './reducers';
+export * from './actions';
